Save RAM and some cycles by storing permutations as ints.
This is a trick based on picking candidate "names" such that
a permutation can be viewed an an C-digit integer in base C,
where C is the number of candidates.

Boosted the largest number of candidates tried from 10 to 11.
This takes a long time to run. Kill the job if you get bored.