Greatly expand the universe of Unicode code points used for names.
There are only about 2 thousand code points that can't be used. Use
all the rest. It's not our job to guess at which are "plausible"
in names. "Be generous in what you accept" is good advice, and not
even Elon Musk could invent a name outside the set we use now ;-)