Do much more testing of very similar score dicts.
It's not really a "strong test" to check that permutations are
evenly distributed across radically different score dicts.

So we only build a brand new score dict once in every 128 tries
now.

Then, 128 times in a row, that dict is mutated by adding
one to the score of a single randomly picked candidate.

So, as close as possible to being identical to the last one
tried.

Testing has gone great through 10 candidates. Checking 11
probably won't finish until tomorrow.